Pulled getBinary to new BinaryController
